refactor: split search query tests
